refactor(factory): reorganize strategy factories
