+++ backend/src/main/scala/gitinsp/infrastructure/factories/ClaudeChatModelFactory.scala
